+++ src/main/java/br/com/pvprojects/restspring3/api/v1/controller/PersonController.java
@@ -9,6 +9,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -62,6 +63,11 @@ public class PersonController {
   @GetMapping(value = "/{id}")
   public PersonModel getPerson(@PathVariable String id) {
     return personService.getPerson(id);
+  }
+
+  @GetMapping(value = "/search/{name}")
+  List<PersonModel> getPersonByName(@PathVariable("name") String name) {
+    return personService.getPersonByName(name);
   }
 
   @DeleteMapping(value = "/{id}")
