+++ src/main/java/br/com/pvprojects/restspring3/util/Util.java
@@ -1,5 +1,7 @@
 package br.com.pvprojects.restspring3.util;
 
+import static org.apache.commons.lang3.StringUtils.EMPTY;
+
 import lombok.experimental.UtilityClass;
 import org.apache.commons.lang3.StringUtils;
 
@@ -10,7 +12,7 @@ public final class Util {
   public static final String TIME_ZONE = "America/Sao_Paulo";
 
   public static String cleanCpf(String cpf) {
-    cpf = cpf.replaceAll("[^0-9]+", "").trim();
+    cpf = cpf.replaceAll("[^0-9]+", EMPTY).trim();
     return cpf;
   }
 
@@ -18,12 +20,12 @@ public final class Util {
     if (!StringUtils.isBlank(string)) {
       return string
           .trim()
-          .replace(" ", "")
-          .replace(",", "")
-          .replace(".", "")
-          .replace("/", "")
-          .replace("_", "")
-          .replace("-", "");
+          .replace(" ", EMPTY)
+          .replace(",", EMPTY)
+          .replace(".", EMPTY)
+          .replace("/", EMPTY)
+          .replace("_", EMPTY)
+          .replace("-", EMPTY);
     }
     return string;
   }
